src/requirements.txt
django
psycopg2-binary
djangorestframework
markdown
django-filter
pygments